send read access token to be used to subscribe to attribute changes (#68)
This will reduce load from the database because when the client queries a list of attribute it will receive an access token for each attribute. When the client then subscribes to the attribute changes, the access token will be send along. The backend will then verify the access token instead of checking the database.